Fixed performance in infection demo!  :)
Changed the 2-bit colour model used previously (there are only
three colours after all) to an 8-bit colour model (which plasma
uses).  JDK 1.1 implemented the 2-bit model efficiently ... but
not, it seems has any JDK since (possibly) 1.3.  They manage
the 8-bit model OK though.

Forgot to mention in the previous commit message: changed the
infect model to do its evolve cycle in parallel (8-way) so as
to take advantage of multicore.  In occam-pi, this would just
be (almost) just the replacing of a SEQ-loop with a PAR-loop.
In JCSP, it's a little harder (because constructing processes
for the Parallel means constructing new objects in Java) ...
but it's not too bad.  Makes a difference - at least, it drives
my dual-core Pentium/Mac flat out, :).