Generics for net2
Changes to net2 and one file in lang.

This commit adds Generics to JSCP networked channels, with a couple of gotchas...

On the implementation side, there are some nasty casts from Object -> T (Which the end-user will not be aware of).

Also, I have not made generic the net2 stuff which is deprecated.

If a Networked receiving channel (net2one or net2any) gets an Object with the wrong type (ie. not T) the user's code will crash with a ClassCastException.
(e.g. your sender is of type NetChannelOutput<Integer> and your recipient of type NetChannelInput<String>, a ClassCastException will be thrown by the recipient when they attempt a read () after the sender writes to the channel.

With luck, this changeset is bug-free and won't impact existing code. I have given it a quick test, but some further testing by others would be nice :-)

tpc